Finally

+++ app/src/main/java/com/example/flappycompose/data/models/Bird.kt
@@ -1,6 +1,6 @@
 package com.example.flappycompose.data.models
 
 data class Bird(
-    val x: Int,
-    val y: Int
+    val x: Float,
+    val y: Float
 )
