+++ app/src/main/java/com/example/flappycompose/data/models/Tube.kt
@@ -1,6 +1,6 @@
 package com.example.flappycompose.data.models
 
 data class Tube(
-    val x: Int,
-    val spaceStartY: Int,
+    val x: Float,
+    val spaceStartY: Float,
 )
